refactor(tree): unify traversal logic and tree structure
