+++ src/mercurytools/tree/binary_search_tree.py
@@ -3,11 +3,43 @@ from ..core.nodes import BinaryTreeNode as Node
 
 
 class BinarySearchTree(TreeBase):
-    def __repr__(self):
-        return f"{self.__class__.__name__}({list(self.inorder())})"
+    def __contains__(self,value):
+        current=self._root
+        while current:
+            if value<current.data:
+                current=current.left
+            elif value>current.data:
+                current=current.right
+            else:
+                return True
+        return False
 
-    def __iter__(self):
-        return self.inorder()
+
+    def _remove(self,node,value):
+        if not node:
+            return node,False
+        if value<node.data:
+            node.left,deleted=self._remove(node.left,value)
+            return node,deleted
+        if value>node.data:
+            node.right,deleted=self._remove(node.right,value)
+            return node,deleted
+        if not node.left and not node.right:
+            return None,True
+        if not node.left:
+            return node.right,True
+        if not node.right:
+            return node.left,True
+        successor=self._min_node(node.right)
+        node.data=successor.data
+        node.right,_=self._remove(node.right,successor.data)
+        return node,True
+
+    def _min_node(self,node):
+        while node.left:
+            node=node.left
+        return node
+
 
     def insert(self,data):
         if not self._root:
@@ -33,17 +65,6 @@ class BinarySearchTree(TreeBase):
             else:
                 return
 
-    def find(self,value):
-        current=self._root
-        while current:
-            if value<current.data:
-                current=current.left
-            elif value>current.data:
-                current=current.right
-            else:
-                return True
-        return False
-
     def remove(self,value):
         self._root,deleted=self._remove(self._root,value)
         if deleted:
@@ -51,52 +72,18 @@ class BinarySearchTree(TreeBase):
             return value
         raise ValueError(f"{value} not found")
 
-    def _remove(self,node,value):
-        if not node:
-            return node,False
-        if value<node.data:
-            node.left,deleted=self._remove(node.left,value)
-            return node,deleted
-        if value>node.data:
-            node.right,deleted=self._remove(node.right,value)
-            return node,deleted
-        if not node.left and not node.right:
-            return None,True
-        if not node.left:
-            return node.right,True
-        if not node.right:
-            return node.left,True
-        successor=self._min(node.right)
-        node.data=successor.data
-        node.right,_=self._remove(node.right,successor.data)
-        return node,True
-
-    def _min(self,node):
+    def min(self):
+        if not self._root:
+            return None
+        node=self._root
         while node.left:
             node=node.left
-        return node
+        return node.data
 
-
-    def inorder(self):
-        def _in(node):
-            if node:
-                yield from _in(node.left)
-                yield node.data
-                yield from _in(node.right)
-        return _in(self._root)
-
-    def preorder(self):
-        def _pre(node):
-            if node:
-                yield node.data
-                yield from _pre(node.left)
-                yield from _pre(node.right)
-        return _pre(self._root)
-
-    def postorder(self):
-        def _post(node):
-            if node:
-                yield from _post(node.left)
-                yield from _post(node.right)
-                yield node.data
-        return _post(self._root)
+    def max(self):
+        if not self._root:
+            return None
+        node=self._root
+        while node.right:
+            node=node.right
+        return node.data
